Fix standalone build issue after upgrading to libspdm 2.1.0.

+++ include/spdmapplib_common.hpp
@@ -184,7 +184,7 @@ bool spdmGetData(spdmItem& spdm, libspdm_data_type_t configType, T& configData,
                  libspdm_data_parameter_t parameter)
 {
     T data;
-    uint32_t data_size;
+    size_t data_size;
 
     data_size = sizeof(data);
     if (!validateSpdmRc(libspdm_get_data(spdm.spdmContext, configType,
